requirements.txt
APScheduler==3.6.1
cacheout==0.11.1
paho-mqtt==1.4.0
requests==2.24.0
pyyaml==5.3.1
xlrd==2.0.1